Add files via upload
at first i made the htlm and css for the first page and the logo and enter button,  after this the "enter.html" part and the second css.
i used flexbox for the whole stuff, and the page a littlebit mobile friendla, but i need to make a few change with the last part cos it doesn't work well. LOL
on the enter.html file there are some tags  what you can use already for the JS part. 
i gave a few id, but i think it wont be enough.
i hope it will be understandable.